Add image building capability
This adds roles for building RDO Manager images.

The image building is done in three phases.
First, a base cloud image is downloaded, gets the
appropriate repos installed, and is updated.
Second, this base image has all packages installed
for the overcloud-full image.
Third, the overcloud-full image gets prepped to boot
as the undercloud, by adding a stack user and hostname.

The base cloud image is fed to DIB for building the deploy
and IPA ramdisks, and the overcloud-base is fed to DIB for
building the overcloud-full image.

These images are then injected into the undercloud.qcow2.