Don't use in memory inventory for ssh config
This led to some odd behavior, where multiple jobs
running concurrently on the same jenkins slave could
end up trying to use the same ssh.config.ansible. This
means they end up deploying to the same undercloud
virtual machine at the same time, which ended rather
poorly.

This seems like an ansible bug, but it is easy enough
to configure the ssh.config in the ansible config
instead of via the in memory inventory.

Change-Id: Iff50ec0b00acb307c72f85efa9295d550ec5ed60