Add fix for RHBZ 1301290 to undercloud image
This is a a one line removal, that is removing something
that is a noop if you use older versions of puppet-mysql,
but which breaks deploying HA if you use a recent version.

It has 3 +2's upstream, but tripleoci is broken again. RDO
should not need to be broken just because tripleoci is.

This patch also removes some workarounds that are no longer
needed.

The only other remaining workarounds are setting SELinux to
permissive on the undercloud, and some hacks on DIB and
python-tripleoclient to make them not put EPEL on the
images.

https://bugzilla.redhat.com/show_bug.cgi?id=1301290
https://bugs.launchpad.net/tripleo/+bug/1537720
https://review.openstack.org/272149

Change-Id: I011a12724f0a15476e6cc355b0827ce8f3ed7774