Dont run CI on changes to image builder role
The current gate job uses pre-built images, so there
is no point in running it against changes only in the
image building role.

It would be good to create a job that basicly runs the
full promote pipeline only on changes to the image
building role, but that will be in a follow up patch.

Change-Id: Ic8f85f564dd77a457f92ee4b8f7e85350d3a4a5b