Remove the need for system-site-packages
In order to get around the behavior described in
https://github.com/ansible/ansible/issues/13773 it
is better to specify the python interpreter to be
the one outside the virtualenv, rather than using
system-site-packages in the virtualenv.

Also, since quickstart.sh is now passing a url for the
undercloud.qcow2, there is no need for multiple
quickstart-*.yml playbooks.

Change-Id: I442dac7dff451fdb9e3b776f969992ff9421cdaa