Update overcloud deploy with upstream fixes
Update workaround for tripleo.sh pingtest
Update to review that allows the pingtest to work for both
old and new heatclients.

Use a single worker overcloud when using the pingtest in
order to improve the reliability.

Don't do introspection on the HA job. This is not needed,
since we run it on the nonHA job. This will save some time
on the longer running job.

Change-Id: I7ba1465f18cdee7980670f5d7049356f37765a9e