Tune default memory and CPU
The defaults are meant to give a working virtual setup that
allows for an HA deployment with a single ceph node and a
single compute node on a 32G RAM virtual host.

However, the HA CI jobs are having issues where undercloud
services are crashing during deploy. Reducing the number of
CPUs on the undercloud will mean fewer workers for the
undercloud services, which will hopefully lead to less
memory being consumed.

Change-Id: Icdbca34edc3b7af350a388ab30a01274b2fdcd08