introduce global "nodes" configuration role
There are several bits of configuration that are duplicated throughout
the tripleo-quickstart project.  This is an attempt to introduce a
little DRY.  This commits creates a "nodes" role that only provides a
`defaults.yml`, which defines pretty much everything we need to know
about nodes and flavors.

Any other role that needs information about the sizing of nodes (for
flavors, deploying libvirt guests, for ironic node profiles, etc)
needs to include the `nodes` role as a dependency.

With this information, we can:

- Change memory for a particular host role with a single variable.
  Want smaller computes?  That's `-e compute_memory=2048`.  Want more
  cpus for you controllers?  `-e controller_vcpu=2`.

- Automatically create the necessary flavors to support the
  deployment.  We ignore the default flavors, and always create a set of
  flavors that match our specific deployment using (all prefixed with
  `oooq_` for `tripleo-quickstart`.

- Automatically apply the correct profile tag to nodes in the
  `instackenv.json` file:

        "capabilities": "profile:{{ node.flavor }},{% endif %}boot_option:local"

It both simplifies the playbooks and makes things easier to configure
at runtime.

Change-Id: Id55e1baa0adf10f704dec2516e98a112be381d14