Refactor and simplify image build
This merges all of the images roles that are used
for building the undercloud.qcow2 into a single
images/build role. This role is then subdivided via
seperate tasks files for each stage of the build.

This (hopefully) makes the structure more clear and
understandable for others to use this image building
process on other base images.

This also allows consolidating all of the default
variables into a single role.

Another major simplification included in this patch is
to no longer rely on python-tripleoclient for image
building. Building images using python-tripleoclient adds
a totally unneeded and overcomplicated abstraction of
diskimage-builder. Instead, we just run diskimage-builder
directly. This will also make it easier to transition to
not using diskimage-builder at all, but instead using
virt-builder.

As part of that transition, I audited all of the elements
which python-tripleoclient included. Many were removed
all together since they only installed packages, and we
are installing packages in a pre-build step. There were
a few elements that had a single line of bash, and those
lines of bash are run using the dib-workaround-default
template in a pre-build step instead. This is just a
first step to make these more visible. All of these should
either be fixed in packaging or dropped.

We also stop building the bash deploy ramdisk in this
patch. It was flaky at best, and nearly impossible to
troubleshoot. The drawback with this, is needing some
hack to make the ironic-python-agent ramdisk work on
liberty. There are two choices:
1. install tgt from EPEL
2. install ironic-python-agent from mitaka delorean
   (this gets us LIO support)

I really don't like 1 because LIO > tgt, but also because
nobody is testing the tgt code path on RHEL based distros.
Further, the stable branch is not gated anywhere in Ironic.

Change-Id: Iefb1589decb9f0e62416409bb912b0a752dfe851