CI playbook and jjb overhaul
playbook overhaul:

1. Reduces repition by using the same playbooks for
liberty and mitaka, with variables being used to get
the same behavior as before. Similarly, these same
playbooks can be used for delorean or official centos
packages.

2. Make the playbooks more straightforward to run
outside of CI, by including the manual provisioner
role. This makes it easier for a developer to quickly
reproduce a CI result locally.

jjb overhaul:

1. We now will only gate on a minimal topology for
patches that touch the libvirt roles. The ha topology
gives too many false negatives to be used reliably for
gating. There is still an ha job for the promote pipeline,
but that job is periodic, so false negatives are not as
costly.

2. We add a gate for building images. This will run a
full image build, and then test the built images using
the minimal topology. This gate will only run on changes
to the image building roles or playbooks.

3. We simplify the instructions posted on the job page
for each job. Previously, these were a direct copy paste
of what CI actually runs. This is not actually very useful
for a developer who wants to reproduce a CI result. She
does not need to be using khaleesi or the hacks that
entails to use both ansible1 and ansible2. Instead, she
would just want to see concise instructions on how to
use tripleo-quickstart to reproduce the job.

Change-Id: I414222aaf2777e078160462761a1b77c23ab56bd