automatically cache undercloud images
with this commit, once an image is fetched successfully and we
validate the checksum we save the image using the md5 checksum as the
name, and in subsequent runs we use the cached copy rather than
fetching it again.

this also means that we support arbitrarily named remote images (so
you can fetch "undercloud-image-with-extra-sauce.qcow2" and it will
work just fine).

Change-Id: I1c7b1c4ba110f2cc730c8a694494772dd52ad487