Pin ansible to 2.0.0
The ansible 2.0.1 release caused all of the gate jobs to
start bleeding together. So 2 jobs started at the same
time would be operating on the same virthost.

It would be a good idea to track down what specifically
is causing this new behavior, but for now we just pin
to a known good release to get the gates back working.

This also changes the default validation to be the ping
test instead of tempest. Tempest is broken on trunk at
the moment, so it does not make sense as the default.

Change-Id: I2e5cd2d99dacf4b306a0b82a57d87ab0800988fa