the "get undercloud vm ip address" task would never fail
because the get-undercloud-ip.sh script always returns success (i.e.,
exits with a return code of 0), the task is never considered to fail,
even if we never get an ip address.

this commit has the script return an exit code based on whether or not
we are able to determine the undercloud, and changes the loop
condition to look at the exit code rather than the output of the
script.

Change-Id: I7daf30576c185c475fe744f6594a50f8e761c4bb