Reduce image build time by 7-8 min
Build ironic-python-agent ramdisk in parallel to overcloud-full.
This uses async tasks as described here:
http://docs.ansible.com/ansible/playbooks_async.html

This change also enables sudo without a tty.  Distros vary, some
have this as the default.  When the ironic image building is happening
as an async task, it'll try (and fail) to sudo without removing
'Defaults requiretty' from /etc/sudoers

Change-Id: I3fc899282420f8c6cc68cecbe8bffa1772948bd8