Use tags to create breakpoints
This change adds all of the existing roles to the
quickstart.yml playbook used by quickstart.sh.

It also adds the ability for quickstart.sh to pass tags
to the ansible-playbook command. The default is now
scenario 6 below. This gives the user of quickstart.sh
some known good scripts to complete the install, but
does not run them.

The tags can be used to get the following scenarios:

1. undercloud booted
  --tags untagged

2. undercloud booted w/ all scripts needed for undercloud
   install, but not run
  --tags "untagged, undercloud-scripts"

3. undercloud install complete
  --tags "untagged, undercloud-scripts, undercloud-install"

4. ready to run deploy
  --tags "untagged, undercloud-scripts,
          undercloud-install, undercloud-post-install"

5. ready to run validation
  --skip-tags "overcloud-validate"

6. undercloud booted w/ all scripts needed for undercloud
   install, deploy, and test, but not run
  --tags "untagged, undercloud-scripts, overcloud-scripts"

7. everything
  --tags all

Change-Id: I94b3c7b6943a3688f82e1a4990c3f277c6c77ce2