make network configuration as easy as node configuration
With this change, the `nodes` role becomes the `common` role and
includes both node and network configuration.  The network handling
has been substantially simplified: we're no longer special-casing the
external network in our tasks, which means we can use a simple
`with_items` loop for creating/destroying networks.

This introduces an `external_network_cidr` setting that can be used to
control the address range, netmask, and dhcp range (sort of) of the
external network with a single config setting (useful if our default
conflicts with an existing network range in the local environment).

this is the non-wip version of 265350.

Change-Id: I49bf4f79ac7b0f7a2ccfeec115a6060595ecf1b7