Add config file support to quickstart.sh
When working on a demo using quickstart.sh, it became
pretty obvious that we need a way to pass an arbitrary
YAML config. Otherwise, quickstart.sh will always give
the same defaults except for the settings we have
directly wired in.

This combined with the CI change to use different config
files for the different jobs, allows us to use
quickstart.sh to run exactly what was run in CI.

The default config file used will be minimal.yml. This
is the most stable job, so would give the best new user
experience.

Change-Id: Ib566cf1bcc75e6e4caec99afee094221dd5be5ad