Refactor CI jobs to be usable via quickstart.sh
It would be nice to reproduce what ran in CI by using
the quickstart.sh tool. This refactors all of the jobs
that don't build images to use the same playbook used
by quickstart.sh by passing a different config file
per job.

We could build upon this by adding more tags and using
the same playbook for image building as well.

Even better would be to run quickstart.sh directly in
CI. However, this change gets us closer to that in the
interim.

This also fixes the minimal topology job to trigger on
changes to common role.

Change-Id: I41161c8f5af0e7f2abd5f85d9abbc48fc8c326d7