fail in a more useful fashion
if the bootstrap fails -- in particular, if "pip install" fails to
install dependencies -- the quickstart.sh script would continue on its
merry way, leading to failures later on.

with this commit, we catch failures during bootstrap and remove the
working directory, which allows us to start cleanly on the next run.

the root cause behind #29 is that gcc needs to be installed on the
host in order to build the pycrypto package.  this change reports the
failure earlier and makes it easier to recover.

Change-Id: Ida1f51582625cd05e3663bc2024415229b1a5ab8