add some documentation
this adds a docs/ directory, into which we move the "contributing"
text from the README, and add docs on running the quickstart
unprivileged and a basic summary of configuration options.

this discards the "image pre fetching" documentation, since that has
largely been automated through the caching introduced in 1edd6cb.

Change-Id: I5d72b5e339478b8df2cbf0012ff7d3493bde8d0a