work around https://github.com/ansible/ansible/issues/14958
If we have a playbook like this:

    - hosts: all
      roles:
        - libvirt/setup

    - hosts: undercloud
      roles:
        - tripleo/undercloud

And we run it like this:

    ansible-playbook -i remotehost, playbook.yml

The the generated ssh config would attempt to use `localhost` as the
proxy host, because Ansible doesn't correctly handle `ansible_host`
when delegating tasks.  This change checks explicitly if
`ansible_host` == `localhost`, and assumes in that situation that
`inventory_hostname` is authoritative.

Change-Id: Ie8d7e0bfbb9157aa7d097a26336bd8f1f87de325