Migrate the base publishing location to defaults.
Putting the base location passed to rsync into the defaults
section allows it to be overridden, so the publishing playbook
can be reused elsewhere by overriding publish_dest_[user|host|basepath]

This change also introduces the ability to pass an optional identity
file that is passed to rsync via publish_dest_key variable.

It also ensures that the destination directory exists, creating it if
needed.

Change-Id: I28c49d970d17dbe4b8cf4b2844736e057da31c3e