Make custom flavors compatible with stock flavors
By using the oooq_ in our capabilities for the ironic
nodes, it makes it impossible to use the stock flavors
created by `openstack undercloud install`. If we instead
just use the flavor name without oooq_, it allows us to
have custom flavors, while still allowing the stock
flavors to work.

Change-Id: I57c45228073892e3bbc228aecd86df4b2abcca02