refactor tags in (over|under)cloud roles
the per-task tagging in the existing roles, combined with our use of
tags as a control structure in quickstart.sh, made it very easy to
make changes that would work fine in a full deploy but would break
with the default quickstart options.

this commit removes all the per-task tags and applies them to include:
stanzas, and moves the tasks themselves into separate files.

Change-Id: Id92429b126dce39ff4ecdef682aad22b002bfa35