remove contents of XDG_RUNTIME_DIR in provision/teardown
bkero's involuntary CI of our teardown tasks has turned up some issues
that only crop up when running the quickstart multiple times on the
same host.

This commit attempts to prevent a number of these problems by ensuring
that the content of the `$XDG_RUNTIME_DIR` directory are removed when
we removed the non-root user account.

Change-Id: I533ae167f4d259f844bfc042f7e15859bbac0681