Move Ironic pxe_ssh driver configuration to libvirt role
We are using qemu://session to manage the overcloud VMs,
which requires a change to the Ironic pxe_ssh driver
configuration. However, by doing this in the
undercloud-install.sh script, we break someone following
the instructions to "Run openstack undercloud install,
then follow upstream instructions."

Instead, we should do this ironic configuration at the end
of the libvirt role, so that the alternative instructions
work again.

Change-Id: I0f94547f0aea251daa6c1ec3a8351c249e69a0a3
Resolves: https://github.com/redhat-openstack/tripleo-quickstart/issues/65