Add -tests packages with Tempest tests to undercloud
Thanks to the Tempest Plugin interface many tests based on Tempest classes
but not in scope for the core Tempest enough have been moved to project
repositories.
The entry points used by Tempest to discover the tests are usually defined
in the -common package with the rest of setuptools entries, but this means
that if the corresponding -tests package is not installed, the Tempest
discovery will fail (entry point without code). So all the -tests packages
which contain Tempest-style tests for packages already part of the undercloud
need to be included as well.

Change-Id: I1f7854b4146029b4826c2700d01dafcf6c825fd5