Use defaults for minimal config
We need to make sure our defaults actually work.
This change removes all configuration from the minimal.yml
config file, with two exceptions.

- We enable introspection.
- We only create the minimum number of nodes.

In order to do this, we need to change the default for
compute_memory, since we can not launch the ping test VM on
a compute node that small.

Change-Id: I668deaacb428db940b35458775dccfd9c8cc461f