add scripts for performing YAML validation
this commit adds scripts for validating YAML documents, as well as a
wrapper for git pre-commit hooks.  The ci-scripts/validate-yaml
script will check individual files or entire directories, e.g:

    ./ci-scripts/validate-yaml playbooks

The ci-scripts/pre-commit-hook script is meant to be installed as a
git pre-commit hook:

    ln -s $PWD/ci-scripts/pre-commit-hook .git/hooks/pre-commit

This will run any executable scripts in `ci-scripts/pre-commit.d`.
There is one scripts there, `ci-scripts/pre-commit.d/validate-yaml-in-commit`, which will validate all of the .yml documents contained in a commit.  This can prevent CI failures do to simple syntax errors by ensuring that the yaml documents are correct before submitting a change to gerrit.

Change-Id: I088e7b56a066e9b8520af9d54489b8a61d8bc29b