Switch to using standalone role for image building
Building an undercloud image was not part of the spec
for importing tripleo-quickstart. However, it is crucial
for the CI that runs in RDO. Eventually, we should move
the playbook responsible for image building out of this
repo as well, but for now we just externalize the role
it uses via ansible-galaxy.

Change-Id: If440d25c0d1687fd9bfc9d78ae35a4547b6866e0