images/build role has moved to its own repo
Image building was not part of the spec to import
tripleo-quickstart. However, RDO relies heavily on
that capability for their CI. We have moved the image
building role to its own repo in the RDO github, and
could easily import it into the OpenStack git if
TripleO decided to use it in the future.

Change-Id: I2deb3f6a7f8d58fc324fa3c9585dd5a844772e4d
Partial-Bug: 1571800