successfully fail when image download fails
it turns out that always: in ansible is not the same as finally: in
Python, in that always: does not propagate failures.  This patches
adds a rescue: clause that sets a variable in the event of a failure
that we can then use to explicitly fail if necessary.

This should probably be considered a temporary solution until we
figure out whether this behavior of always: in ansible is a bug or is
expected.  The current docs are unclear on that point.

Change-Id: I16f42b0224a487a6de36d86139fe8265b33daec9
Closes-bug: #1575234