Collect logs from failed heat software deployments
Currently, if the overcloud fails to deploy due to
an issue with on of the heat software deployments, we
do not get any useful logs. This change pulls that
information from heat and puts it in the stack user's
home directory on the undercloud. CI already grabs any
files ending in .log from there.

This also gets rid of the overcloud-status template,
because it is not being used anywhere.

Change-Id: Ief3364814d17d78cd3a40a7c0159ca0066b5ec6b
Partial-Bug: 1571046