work around ansible issue #15744
with Gabriele's upcoming patches that allow the quickstart to run
without performing teardown operations, we hit Ansible issue 15744
(https://github.com/ansible/ansible/issues/15744), in which Ansible
will incorrectly assume it has gathered facts for a host if it *skips*
a play that would otherwise target that host.

This commit removes `gathering=smart` from our ansible.cfg.  We can
restore that configuration if we pin Ansible to 2.1 or later at some
point in the future.

Change-Id: Idda5289a6ea4c724ad2e936b6feb1128925d0e6f