fix network configuration
When deploying with `network_isolation: true`, we are failing to pass
in the network settings from our configuration into the tripleo
deployment process, leading ultimately to deployment failures when
attempting to use non-default values.

additionally, setting `network_isolation: true` required adding a
number of explicit arguments to `extra_args`, so it was not possible
to control the use of network isolation simply by toggling the value
of `network_isolation`.

This commits allows the deployment to succeed with `network_isolation`
either disabled or enabled, and allows the use of non-default values
for the network configuration.

Change-Id: I68f5a87e390aa2ea8d8c5631eba17dc5dd44270e
Closes-bug: #1581659