Add CI scripts for full deploy and images jobs
This simply copies what we have in JJB into in-tree scripts.
After this merges, we can change the JJB to consume these
scripts. This makes it much easier to change what is run in
CI, since we can get actual feedback on a patch that changes
CI.

It also combines promote and gate jobs into a single script
each. This will dramatically simplify the CI JJB.

Change-Id: I9600c773b3c7194d525b35b4acb502c5eea15323