Remove dependency on the khaleesi project
The khaleesi project is unlikely to be ported to
Ansible2, instead the functionality is being ported
into atomic roles for easier consumption.

We were only using khaleesi to get nodes from centos-ci.

This commit also makes --bootstrap a bit more useful.
Now it will exit before running any ansible, so it can
be used to setup the virtualenv for CI, as well as in
the instructions for reproducing CI.

Change-Id: I63be183f811071a0de61e27668d623d06668040b