Do not archive root directory
DIB is putting image cache in the root directory for the image
building jobs, which then gets uploaded to the artifacts server
for every job. That is 1.5GB of images getting put on the artifacts
server for every job that builds images. Since there is not anything
useful in the root directory anyways it is easiest to just not
archive it.

Change-Id: I4e2f05fbce4ae6dc59e07b828e35edad5f9aafe2