move ironic.conf modifcation to an undercloud role
When running with virthost == localhost the ssh.config.ansible
is not valid as it's trying to proxy through the localhost.
This is fixed in the ansible-role-tripleo-inventory role.
Changing the ironic.conf file on the undercloud must come
after there is a valid ssh config file in all scenarios.
It's safe to move this to the first step in the tripleo/undercloud
role.

Change-Id: I6a5d692a1b1783130fcd3aec2fba9be453fc8e3b