Move default stopping point to just before overcloud deploy
For both the developer and new user use cases, there is rarely a
need to do anything before the undercloud is installed and images
and nodes are uploaded and registered. This patch moves up the
default stopping point to just before the overcloud deploy command
is run. This behavior is still changeable by changing the tags
passed to quickstart.sh.

Also fixes a small UI bug, where usage instructions get print out
even when not using quickstart's default playbook. These instructions
are unlikely to be accurate when used with a different playbook.

Change-Id: I40eb8560732f266a1cd4ac37187e5fbb9b620b8e
Closes-Bug: 1569477