check the overcloud_deployment_result variable to determine playbook exit
Requires: https://review.gerrithub.io/281750
This change reads in the variable set "overcloud_deployment_result" and
determines whether or not to exit the playbook based on the variable result
vs. exiting the playbook immediately after a deployment fails.

This allows us to run inventory on the overcloud nodes, and collect logs
for further debug.  I expect this to transform into it's own role so we
can make adjustments to the level of debugging for a deployment.

Change-Id: I7a51534fd252b6b6390a7c0d9bc6c6cd024b5c24