Move THT clone to undercloud post install
By moving the THT clone to the undercloud post install script, it
allows for a workflow where ansible stops just before we deploy the
overcloud, and then we can hack on the checked out copy of the
templates before running the overcloud-deploy script.

This also adds the ability to pass in an open gerrit review for THT
via a refspec (overcloud_templates_refspec), and the THT checkout will
contain that review.

Change-Id: Idbbd28378db1c06306f70b154bfa81b3fdd5f731