Bump default undercloud vCPUs to 8
Three out of the four chassis' in CentOS CI have slower CPUs, but
they have more of cores. I have tested bumping the number of CPUs
on the undercloud on these slower chassis' and it made a big
difference. More so, the memory pressure on the undercloud is
actually not bad now, so there is not too much risk in hitting
issues where more CPUs causes high memory consumption.

This also increases the default in the HA configuration by the same
factor (2x).

Change-Id: I44f34827774353f9f70e7fd51c524567914059ab