oooq support check, exit on anything other than CentOS and RHEL
At the moment our oooq CI only runs on CentOS and RHEL.  To fully
support a linux distribution we need to have CI running it all the time.
This will exit the playbook with a support message if CentOS or RHEL
are not the distribution of the virthost.

Allow a user to override this check by setting
"quickstart.sh -e supported_distro_check=false"

Change-Id: I71ce50072bf8df6b0ec773e5ab69c2a27d1694a9