Add tox -elinters
OpenStack ansible projects have a series of gate jobs to help with
test coverage.  This patch adds some of the basic frame work needed
for that.   With this path we'll run both flake8 and ansible-lint on
tripleo-quickstart.

Currently, tox -elinters will fail, and this is expected. I plan to
add a series of follow up patches to fix this.  I also plan to enable
check / gate jobs to ensure new patches to tripleo-quickstart are
formatted properly.

Change-Id: Ie2c4efcc7581046fd5d92b38e7b24b36c460c58a
Signed-off-by: Paul Belanger <pabelanger@redhat.com>