fix teardown (tagging and ansible_user)
This patch fixes two large problems with our teardown logic:

(a) the tags applied to the teardown plays combined with the way we
were calling ansible in quickstart.sh meant that we never executed the
teardown tasks with the default `OPT_TEARDOWN=nodes` behavior.

(b) We were calling all of the teardown tasks at the same point, which
meant that the environment and node teardown tasks were being called
with `ansible_user=root`.  This meant that the node teardown in
particular was a no-op, because the libvirt environment of the stack
user was invisible to the tasks running as root.

With this change in place, I can successfully run quickstart.sh
several times in a row without error (previously this would fail
because the teardown tasks wouldn't execute, causing conflicts)

Change-Id: Ic94241c1d87b2dae335c6cb6627a4fdb5419482e
Closes-bug: 1610677
Closes-bug: 1610537