Relocate changes to staged publishing.
The following commit was made to ansible-role-tripleo-image-build:

https://github.com/redhat-openstack/ansible-role-tripleo-image-build/commit/fe1afa02ff7034b5b23eb73d231cd4fc0cee86d4

The current tripleo-quickstart build gates still use this file
for actual publishing.  Over the summer when we decoupled the
build role from this one (tripleo-quickstart proper) the cores
for oooq did not want to locate any publishing code not used by
RDO here, nor did they want to allow this to migrate to the
image build role.  Rather than create a third role (just to do
a file copy) we settled upon having oooq handle it's own file
copy (here), and the rest of the consumers of the now-free-
standing build role could use things there.  This caused duplication.

This "staged" file copy mode used by RDO specifically is slated
to be removed by the following when it lands, however it remains
my opinion that we should migrate this remaining morsel back to
the build role itself.

In order to unblock the current work (below, new rdo pipeline)
however we need these changes where they are being invoked by
gates and such.

Change that will remove staged publish from artib unless we want
to keep it there and remove it from oooq

Related-To: https://review.gerrithub.io/#/c/277420/

(already merged to artib)
Related-To: https://review.gerrithub.io/#/c/287153/

^ was made to support these:
Related-To: https://review.gerrithub.io/#/c/287144/
Related-To: https://review.openstack.org/354365

Change-Id: I9d1fbbd4e73ffe877edb279a17ea897d71922e4c