newton/master add ssl option as default to True for the undercloud
SSL is a default option in upstream.  RDO is missing bugs
because we are not testing w/ SSL e.g. [1].

[1] https://bugs.launchpad.net/tripleo/+bug/1613088

This also sets defaults for the undercloud's public VIP and admin
VIP, since HAProxy will bind to those when we enable the undercloud,
so they need to be values according to the network that was
configured, unless the deployer has set them specifically.

Co-Authored-By: Juan Antonio Osorio Robles <jaosorior@redhat.com>
Change-Id: Ia0ce1a53411a435d25289ee492c794d5a4368149