Add objectstorage flavor and extra blockdevices option
This patch adds an option to create three additional blockdevices per
node. The devices are created as sparse files, therefore not using space
before being used. The option can be used in any flavor, and a new
flavor for Swift has been added that makes use of it.  It is especially
useful to test Swift and other storage setups; and by using three
blockdevices per node one can even test a triple-replicated setup with a
single object storage node.

Change-Id: I3960065916b93bb7616d509e7b72c579f7362efe