Allow ssh to overcloud nodes in inventory
Right now if you attempt to perform a task on
an overcloud node in a oooq playbook the ssh
connection fails because the inventory entry
does not add the ssh_extra_args command pointing
at the ssh config file.

This corrects that and allows tasks to be run
against the overcloud nodes.

Change-Id: I05caa7bb325fc7cc7cddb43e5781b5336972e35d