Make undercloud service workers configurable
This enables us to configure the number of workers for the
undercloud services; which is useful for when resources are limited.

This is configured via a variable called 'undercloud_workers' which
can set the workers for all the services. And each individual service
can be configured too. The pattern is
'undercloud_<service name>_workers'.

Change-Id: I9d381081216cae366e7d2c2896c004ff002ef797