create release files for consistent and current-tripleo
In ci.centos there are errors on missing release files for the master
and current-tripleo promotion jobs.

reference:
tripleo-quickstart-promote-master-delorean-minimal/563
tripleo-quickstart-promote-master-current-tripleo-delorean-minimal/12

error:
master-current-tripleo.yml' does not exist, or is not readable
master-consistent.yml' does not exist, or is not readable

This fixes the release file names to match the repurpose of the
$LOCATION variable introduced in https://review.gerrithub.io/289818

During the promotion full-deploy.sh will use the proper release files
now pointing to the testing versions of either the consistent or the
current-tripleo images while master will still point to the latest
promoted current-tripleo image.

The $LOCATION variable is now a mandatory parameter for promotion jobs.

Co-Authored-By: Attila Darazs <adarazs@redhat.com>
Change-Id: If893fa4c1cd30980b4c6b840e8459a317ad1a17c