Add ability to use overcloud-full.qcow2 for undercloud image
This change adds a "overcloud_as_undercloud" boolean configuration
option which defaults to false, and is set to true in the devmode
configuration file.

When this option is set to true, the undercloud image is not used.
Instead, the overcloud-full.qcow2 is converted into a suitable
undercloud image.

The libvirt/setup/undercloud role is also cleaned up a bit as
part of this change.

Change-Id: Ib6b0e3beb41bd140c1bd327c994fb4c805ae3022