Add all the gating functionality to full-deploy.sh
This parts lived in roles-deploy.sh which should retire now that roles
are completely outside of oooq but also integral part of it as 'extras'.

It also adds a proper call for the new dev mode release when running
dlrn-gate.

To avoid CI errors removal of roles-deploy.sh will be a separate change
after we altered the JJB files to use full-deploy.sh.

Change-Id: Ib9bc1bfcf2c034f4c1083abb0287dd6c8a3f2203