Use the proper private keys for ssh config file
When regenerating the ssh config through the tripleo-inventory role the
private keys were statically defined instead of taken from the
inventory. This can cause errors, as we might have different keys in use
during execution.

For example using the id_rsa_virt_power key was incorrect, as the
provision/remote role adds and defines the correct virthost key as

virt_host_key: "{{ local_working_dir }}/id_rsa_virt_host"

It was still working, but there was a discrepancy between the inventory
and the ssh config.

Change-Id: Ied20e2306fbc1fb94a4654d0eed9d5058bc579ae