Switch default image location back to CentOS CDN
Since changing the way we do the image promotion to make it an
atomic operation, we are no longer seeing the md5sum mismatch
issue. As such, we should return to using the CDN for the default
image location as this is much faster than our file server on OS1.

Change-Id: I1d7e6f2f16c727ab16999a8ac5f0a4a3201c7de8