Return to using ping test in minimal jobs
Unfortunately, the CI infrastructure is not able to handle running
tempest on top of our virtual cloud. We need the minimal job to
have a very low rate of false negatives, and right now it is at
100 percent. As such, we should return to using the ping test until
tempest can be proven to work on quickstart deployed on the
ci.centos infrastructure.

Change-Id: I6f8c41b30d2b6d38d84711de6d73418959ccd580