Properly reload kvm module when trying to set up nested virtualization
The kvm_{intel|amd} kernel module is loaded automatically at boot when
possible/available, regardless of whether qemu-kvm has been installed
or not.
This means we need to rmmod kvm_{intel|amd}, do the nested virt
configuration and then modprobe again.

There is also no need to configure the module to load at boot since
it loads automatically when available.

This is a backported refactored implementation from WeIRDO [1] that
has been proven to work successfully with Packstack and
puppet-openstack-integration.

[1]: https://review.rdoproject.org/r/#/c/3210/

Change-Id: If0b733dcb0f77e6004b133bf4d9a3b66712c63e8