Temporarily stop running tempest in default configs
The main place we test tripleo-quickstart (ci.centos.org), struggles
to provide consistent results running tempest. The hardware there is
heterogeneous, specifically three of the four chassis there have
slower CPU. When a job that runs tempest is scheduled to one of these
nodes, the deployed cloud can not handle the load and fails random
tests. This is problematic because it gives us a 75% false negative
rate for jobs that run tempest.

There is some new hardware being purchased for ci.centos.org that
will be put in a pool that can be specifically scheduled to. In the
worst case, we can re-enable tempest jobs at that time.

There is also some possibility of tuning the performance on the
underperforming CI nodes in order to drop the false negative rate.

Change-Id: I6302d20f869dfcf9517965468752d483c8acafbb