Configuring vcpus for undercloud, controller and compute nodes
By default, undercloud is created with 8 vcpus, while compute and
controller are created only with one. This make the controller
struggle with requests from tempest that runs on undercloud, and
have the concurrency set to vcpus/2, which means 4 threads doing
a lot of requests to controller and this one have only one vcpu
to handle these requests, plus all the services running on the vm.
In my tests, 2 vcpus were enough to make a full tempest run pass
without bottlenecks.
Also keeping the same configuration for HA, since in this case
the workload is spread, so there's no need to change the actual
setup for now

Change-Id: Idbdff31e77b239a4c5ec5e7bf918b89a0a6cf533