VirtualBMC support for tripleo-quickstart
This patch is replacing the usage of the pxe_ssh driver in Ironic in
favor of pxe_ipmitool + VirtualBMC*. The main reason for this patch is
because the pxe_ssh driver is already marked as unsupported and will be
removed from Ironic upstream at the end of the Ocata release.

Older releases: Liberty, Mitaka and Newton will continue to rely on the
pxe_ssh driver.

* VirtualBMC is a proxy that converts IPMI commands to libvirt calls

Partial-Bug: #1645733
Change-Id: Iaaa9fc1fd593cbce045456f4341c461a8eea82ad