Fix devmode for change to TripleO produced images
With the TripleO promotion on Nov. 22, we no longer have any repos
on the overcloud-full image besides the CentOS base repos. This
meant when we tried to install python-tripleoclient in the script
that converts that image to an undercloud image it would fail.

This patch splits up the repo injection tasks from the image
update tasks, and puts the repo injection tasks before the image
conversion.

It also addresses an issue where just after TripleO promotion,
devmode is broken until the buildlogs repo mirror catches up. This
can be up to 2 hours. We fix it by allowing the install of tempest
and test packages to fail without failing the convert script. This
means in that 2 hour window where buildlogs is behind the image
content, we won't have tempest on the undercloud, but that is
better than not being able to boot an undercloud at all.

Finally, it splits out the image injection task, so that it is
always run when using an overcloud as an undercloud, and not
solely for the devmode use case.

Change-Id: Ide7e3814809c54ee57225db848e014f482a60ec7