Add support for external network bridges
Previously NAT was required since bridges can't have addresses,
and the instackenv.json template was expecting an address on the
external bridge to populate pm_addr.

If the external network is designated as a bridge, another network
must be set to NAT in order for the virthost to be able to access
the undercloud, since otherwise there's no address given to
the VM. If no such network is defined, the default libvirt network
(virbr0) is used.

In addition, the requirement to have networks named 'external'
and 'overcloud' has been removed. The first bridge network specified
in the network list will be used as a pxe network regardless of name.

Change-Id: I77f2df7862a71d9a62e4382138bc5287d0687569