Add support for virtual port types
the virtualport type attribute can be used to connect to bridges
that are provided by OpenVSwitch, as opposed to the default
linuxbridge. This functionality is not available to unprivileged
VMs, and will require additional configuration to enable.

The OVS bridges are created in the environment role as part of
setup and teardown, using the ansible ovs module.

Change-Id: I5a569c34a66b67877f7b471d3864e9ae6806bef6
Signed-off-by: Michael Chapman <woppin@gmail.com>