Add ssh tunneling for TripleO-UI
Add a systemd service that tunnels traffic between the virthost and the
undercloud.  We forward ports for the UI itself, and a few other Openstack
services, like Keystone, Heat, Mistral, etc.

This requires a change that sets up the tripleo-ui javascript
config with the proper settings and restarts the httpd service
https://gist.github.com/weshayutin/0af348fafd70cb56c96d6856f073d79b
Completed this change in Ie9d3e7d12c420b9b65c25692435c0457f1233a27

Access to the TripleO-UI is available from http://virthost:3000

Author:    Honza Pokorny <honza@redhat.com>
Co-Authored-By: Wes Hayutin <weshayutin@gmail.com>

Change-Id: Id53c5bd7f199c64bc9cb37df7e63f27c4673f415